move FacetDefinition to @gen3/frontend
