+++ packages/core/src/features/filters/types.ts
@@ -1,5 +1,3 @@
-import { IndexAndField } from '../guppy';
-
 export interface Intersection {
   operator: 'and';
   operands: ReadonlyArray<Operation>;
@@ -187,34 +185,9 @@ export type HistogramBucket = {
   count: number;
 };
 
-export type FacetType =
-  | 'enum'
-  | 'exact'
-  | 'range'
-  | 'age'
-  | 'year'
-  | 'years'
-  | 'days'
-  | 'percent'
-  | 'datetime'
-  | 'toggle' // Note these support alternative UIs
-  | 'multiselect';
-
 export interface AllowableRange {
   readonly minimum?: number;
   readonly maximum?: number;
-}
-
-export interface FacetDefinition {
-  readonly description?: string; //description from _mapping
-  readonly field: string; // full name of field
-  readonly dataField: string; //
-  readonly index: string; // what dataType is this facet for
-  readonly type: FacetType; // classified type based on type + name: e.g. age, year, enumeration, etc
-  readonly range?: AllowableRange; // range of value types
-  readonly hasData?: boolean; // does this facet have data
-  readonly label?: string; // label for facet
-  readonly sharedWithIndices?: Array<IndexAndField>; // if this filter is denormalized across indices
 }
 
 export type IndexedFilterSet = Record<string, FilterSet>;
