+++ packages/core/tsconfig.json
@@ -7,6 +7,7 @@
     "src/**/*.spec.ts*"
   ],
   "compilerOptions": {
+    "rootDir": "src",
     "outDir": "dist/dts",
     "jsx": "react",
     "allowSyntheticDefaultImports": true,
