+++ packages/frontend/src/features/CohortDiscovery/FilterPanels/EnumFacetPanel.tsx
@@ -3,10 +3,11 @@ import { Text, Tooltip } from '@mantine/core';
 import {
   EnumFacetDataChangedFunction,
   FacetDataHooks,
+  FacetDefinition,
   GetEnumFacetDataFunction,
 } from '../../../components/facets';
 import { createChart } from '../../../components/charts/createChart';
-import { FacetDefinition, fieldNameToTitle, HistogramData } from '@gen3/core';
+import { fieldNameToTitle, HistogramData } from '@gen3/core';
 import FacetEnumList from '../../../components/facets/FacetEnumList';
 import { FacetIconButton } from '../../../components/facets/components';
 import { FaUndo as UndoIcon } from 'react-icons/fa';
