+++ packages/frontend/src/features/CohortDiscovery/FilterPanels/createFacetPanel.tsx
@@ -1,7 +1,6 @@
-import React from 'react';
-import { FacetDefinition } from '@gen3/core';
+import React, { ReactElement } from 'react';
+import { FacetDefinition } from '../../../components/facets/types';
 import EnumFacetPanel, { EnumFacetPanelDataHooks } from './EnumFacetPanel';
-import { ReactElement } from 'react';
 
 export const createFacetPanel = (
   facet: FacetDefinition,
